Prevent beeps in property editor
Clicking focus into a property editor which both has
proposal support as well as has a current value which
is not a proposal match (for example, a TextView's text
property can have a hardcoded string which is not a
match for the proposals which are for @-resources) would
cause an audible beep. This CL prevents that.

Change-Id: I0a006f594266cce248b91d065f4d438659b1743b